Keep numeric error code in query errors
When query commands (specifically, nextObject) return an error as a
result document of the form {$err: "message", code: 1234}, we should
allow the application to see the code (and any other fields on the
error).  This is consistent with what we do if the error is returned in
the "errmsg" field on the next line.